Positioned panel correctly.
I believe that the Window padding / View margin cannot be set to zero
(lots on this on qt5 forums). As a result,
  - the QGraphicsView does not fit in the Window if they are given the
    same dimensions;
  - the QGraphicsView is positioned one pixel to far to the right and
    two pixels too low down.

The correct positioning of the QGraphicsView is achieved by making the
Window a bit oversize and letting it go outside the screen.

Under openbox, the "WindowFlag" Qt::X11BypassWindowManagerHint needed
to be set to achieve this.